refactor and some work (WIP)
