+++ motors/pro/MotionMagicTalonFXPro.java
@@ -1,11 +1,8 @@
 package frc.robot.ShamLib.motors.pro;
 
-import com.ctre.phoenix.motorcontrol.TalonFXControlMode;
 import com.ctre.phoenixpro.configs.MotionMagicConfigs;
 import com.ctre.phoenixpro.configs.TalonFXConfiguration;
 import com.ctre.phoenixpro.controls.MotionMagicVoltage;
-import frc.robot.ShamLib.motors.PIDFGains;
-import frc.robot.ShamLib.motors.PIDSVGains;
 
 public class MotionMagicTalonFXPro extends EnhancedTalonFXPro {
     private int kTimeoutMs = 30;
