+++ swerve/ModuleInfo.java
@@ -14,6 +14,7 @@ public class ModuleInfo {
     final boolean driveInverted;
     final boolean turnInverted;
 
+    //TODO: Properly support both phoenix pro and phoenix V5
     private static double mk3TurnRatio  =
             (1.0 / 2048) *
             (1.0 / 12.8) * //
@@ -21,7 +22,7 @@ public class ModuleInfo {
     ;
 
     private static double mk4iTurnRatio =
-        (1.0 / 2048) * //Motor revs
+//        (1.0 / 2048) * //Motor revs
         (7.0 / 150.0) * //Output revs
         360 //Output degrees
     ;
@@ -56,7 +57,7 @@ public class ModuleInfo {
     public static ModuleInfo getMK4IL1Module(int driveMotorID, int turnMotorID, int encoderID, double encoderOffset, Translation2d offset, boolean driveInverted) {
         return new ModuleInfo(driveMotorID, turnMotorID, encoderID, encoderOffset, offset,
                 mk4iTurnRatio,
-                (1.0 / 2048) * //Motor revs
+//                (1.0 / 2048) * //Motor revs
                 (1.0 / 8.14) * //Output revs
                 mk4iWheelCircumference, //Output meters
                 driveInverted,
